requirements.txt
goblet-gcp==0.9.2